est-sale-backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-jose[cryptography]==3.3.0
python-keycloak==2.14.0
cassandra-driver==3.28.0
httpx==0.24.1
python-multipart==0.0.6